+++ breeze-core/src/main/kotlin/com/windea/breezeframework/core/annotations/api/StableApi.kt
@@ -1,5 +1,5 @@
 package com.windea.breezeframework.core.annotations.api
 
-/**暂不稳定的api。*/
+/**稳定的api。*/
 @MustBeDocumented
-annotation class UnstableApi
+annotation class StableApi