+++ breeze-core/src/test/kotlin/com/windea/breezeframework/core/extensions/BooleanExtensionsKtTest.kt
@@ -1,0 +1,31 @@
+package com.windea.breezeframework.core.extensions
+
+import org.junit.*
+
+class BooleanExtensionsKtTest {
+	//IGNORE TESTS
+	
+	@Test
+	fun toInt() {
+	}
+	
+	@Test
+	fun toFloat() {
+	}
+	
+	@Test
+	fun toDouble() {
+	}
+	
+	@Test
+	fun toOpsInt() {
+	}
+	
+	@Test
+	fun toOpsFloat() {
+	}
+	
+	@Test
+	fun toOpsDouble() {
+	}
+}