+++ breeze-core/src/test/kotlin/com/windea/breezeframework/core/extensions/CollectionExtensionsKtTest.kt
@@ -1,0 +1,12 @@
+package com.windea.breezeframework.core.extensions
+
+import org.junit.*
+
+class CollectionExtensionsKtTest {
+	@Test //TESTED
+	fun repeatExtensionTest() {
+		println(listOf(1, 2, 3).repeat(3))
+		println(listOf(1, 2, 3).flatRepeat(3))
+		println(listOf(1, 2, 3).repeatChunked(3))
+	}
+}