+++ CHANGELOG.md
@@ -33,7 +33,9 @@
 * [X] breeze-data 编写`DataSerializer`的相关接口和类，未实现。
 * [X] breeze-core 扩展`LetterCase`，提供更多的显示格式。
 * [X] 补充api注解。
-* [ ] breeze-core 进一步完善`LetterCase`，通过添加泛用显示格式放松限制。
+* [X] breeze-core 进一步完善`LetterCase`，通过添加泛用显示格式放松限制。
 * [ ] breeze-data 代理实现`DataSerializer`的相关类。
 * [ ] DELAY breeze-text en 序数与数量的转化。
 * [ ] DELAY breeze-text en 单数与复数的转化。
+
+hyphen-Words--123